Fix issue: Add a new image if product image is null

+++ backend/src/main/java/com/hugorithm/hopfencraft/service/ProductService.java
@@ -259,22 +259,31 @@ public class ProductService {
             Product product = productRepository.findById(productId)
                     .orElseThrow(() -> new ProductNotFoundException("Product not found with id: %s", productId));
 
-            File oldFile = new File(product.getImage().getPath());
-            if (oldFile.exists()) {
-                if (oldFile.delete()) {
-                    LOGGER.info("Old file deleted successfully");
-                } else {
-                    LOGGER.error("Failed to delete old file");
-                }
-            }
-
             UUID uuid = UUID.randomUUID();
             String filePath = FOLDER_PATH + uuid + "_" + file.getOriginalFilename();
 
-            ProductImage image = product.getImage();
-            image.setName(uuid + "_" + file.getOriginalFilename());
-            image.setType(file.getContentType());
-            image.setPath(filePath);
+            if (product.getImage() == null) {
+                ProductImage image = new ProductImage(
+                        uuid + "_" + file.getOriginalFilename(),
+                        file.getContentType(),
+                        filePath
+                );
+                product.setImage(image);
+            } else {
+                File oldFile = new File(product.getImage().getPath());
+                if (oldFile.exists()) {
+                    if (oldFile.delete()) {
+                        LOGGER.info("Old file deleted successfully");
+                    } else {
+                        LOGGER.error("Failed to delete old file");
+                    }
+                }
+                ProductImage image = product.getImage();
+                image.setName(uuid + "_" + file.getOriginalFilename());
+                image.setType(file.getContentType());
+                image.setPath(filePath);
+
+            }
 
             file.transferTo(new File(filePath));
             Product p = productRepository.save(product);
